+++ .github/workflows/test-pro.yaml
@@ -43,20 +43,5 @@ jobs:
       run: |
         brew install make automake libtool
 
-    # When we build all dependencies with the 11.0 target, we can downgrade this back.
-    - if: ${{ matrix.os == 'macos-14' }}
-      run: echo "MACOSX_DEPLOYMENT_TARGET=14.0" >> "$GITHUB_ENV"
-
-    - if: ${{ matrix.os == 'macos-13' }}
-      run: echo "MACOSX_DEPLOYMENT_TARGET=13.0" >> "$GITHUB_ENV"
-
     - name: Build and test wheels
-      uses: pypa/cibuildwheel@v3.1.3
-
-    # - name: Setup tmate session
-    #   uses: mxschmitt/action-tmate@v3
-
-    - uses: actions/upload-artifact@v4  # https://github.com/actions/upload-artifact/issues/478
-      with:
-        name: curl_cffi-${{ github.sha }}-${{ matrix.os }}.zip
-        path: ./wheelhouse/*.whl
+      uses: pypa/cibuildwheel@v3.3.1
