+++ AGENTS.md
@@ -1,6 +1,8 @@
 # curl_cffi
 
 ## Setup commands
+
+
 - stable: `pip install curl_cffi`
 - beta: `pip install curl_cffi --pre`
  
@@ -9,6 +11,10 @@
 - Current supported version: Python 3.10 and above, do not use Python 3.9 syntax.
 
 ## Dev environment tips
+
+We use conda to create and manage virtual environment, to activate:
+
+    conda activate curl_cffi
 
 ## Testing instructions
