+++ curl_cffi/requests/impersonate.py
@@ -1,6 +1,6 @@
 import warnings
 from dataclasses import dataclass
-from enum import Enum
+from enum import Enum, IntEnum
 from typing import Literal, Optional, TypedDict
 
 
@@ -195,6 +195,56 @@ class ExtraFpDict(TypedDict, total=False):
     http2_no_priority: bool
 
 
+class CipherName(IntEnum):
+    TLS_RSA_WITH_3DES_EDE_CBC_SHA = 0x000A
+    TLS_RSA_WITH_AES_128_CBC_SHA = 0x002F
+    TLS_DHE_RSA_WITH_AES_128_CBC_SHA = 0x0033
+    TLS_RSA_WITH_AES_256_CBC_SHA = 0x0035
+    TLS_DHE_RSA_WITH_AES_256_CBC_SHA = 0x0039
+    TLS_RSA_WITH_AES_128_CBC_SHA256 = 0x003C
+    TLS_RSA_WITH_AES_256_CBC_SHA256 = 0x003D
+    TLS_DHE_RSA_WITH_AES_128_CBC_SHA256 = 0x0067
+    TLS_DHE_RSA_WITH_AES_256_CBC_SHA256 = 0x006B
+    TLS_PSK_WITH_AES_128_CBC_SHA = 0x008C
+    TLS_PSK_WITH_AES_256_CBC_SHA = 0x008D
+    TLS_RSA_WITH_AES_128_GCM_SHA256 = 0x009C
+    TLS_RSA_WITH_AES_256_GCM_SHA384 = 0x009D
+    TLS_DHE_RSA_WITH_AES_128_GCM_SHA256 = 0x009E
+    TLS_DHE_RSA_WITH_AES_256_GCM_SHA384 = 0x009F
+    TLS_AES_128_GCM_SHA256 = 0x1301
+    TLS_AES_256_GCM_SHA384 = 0x1302
+    TLS_CHACHA20_POLY1305_SHA256 = 0x1303
+    TLS_ECDHE_ECDSA_WITH_3DES_EDE_CBC_SHA = 0xC008
+    TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA = 0xC009
+    TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA = 0xC00A
+    TLS_ECDHE_RSA_WITH_3DES_EDE_CBC_SHA = 0xC012
+    TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA = 0xC013
+    TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA = 0xC014
+    TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA256 = 0xC023
+    TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA384 = 0xC024
+    TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256 = 0xC027
+    TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA384 = 0xC028
+    TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256 = 0xC02B
+    TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384 = 0xC02C
+    TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256 = 0xC02F
+    TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384 = 0xC030
+    TLS_ECDHE_PSK_WITH_AES_128_CBC_SHA = 0xC035
+    TLS_ECDHE_PSK_WITH_AES_256_CBC_SHA = 0xC036
+    TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256 = 0xCCA8
+    TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256 = 0xCCA9
+    TLS_ECDHE_PSK_WITH_CHACHA20_POLY1305_SHA256 = 0xCCAC
+
+    @classmethod
+    def get(cls, cipher_id: int) -> str:
+        try:
+            return cls(cipher_id).name
+        except ValueError:
+            return ""
+
+
+cipher_name = CipherName
+
+
 # TLS version are in the format of 0xAABB, where AA is major version and BB is minor
 # version. As of today, the major version is always 03.
 TLS_VERSION_MAP = {
@@ -207,45 +257,7 @@ TLS_VERSION_MAP = {
 
 # http://www.iana.org/assignments/tls-parameters/tls-parameters.xml
 # via BoringSSL
-TLS_CIPHER_NAME_MAP = {
-    0x000A: "TLS_RSA_WITH_3DES_EDE_CBC_SHA",
-    0x002F: "TLS_RSA_WITH_AES_128_CBC_SHA",
-    0x0033: "TLS_DHE_RSA_WITH_AES_128_CBC_SHA",
-    0x0035: "TLS_RSA_WITH_AES_256_CBC_SHA",
-    0x0039: "TLS_DHE_RSA_WITH_AES_256_CBC_SHA",
-    0x003C: "TLS_RSA_WITH_AES_128_CBC_SHA256",
-    0x003D: "TLS_RSA_WITH_AES_256_CBC_SHA256",
-    0x0067: "TLS_DHE_RSA_WITH_AES_128_CBC_SHA256",
-    0x006B: "TLS_DHE_RSA_WITH_AES_256_CBC_SHA256",
-    0x008C: "TLS_PSK_WITH_AES_128_CBC_SHA",
-    0x008D: "TLS_PSK_WITH_AES_256_CBC_SHA",
-    0x009C: "TLS_RSA_WITH_AES_128_GCM_SHA256",
-    0x009D: "TLS_RSA_WITH_AES_256_GCM_SHA384",
-    0x009E: "TLS_DHE_RSA_WITH_AES_128_GCM_SHA256",
-    0x009F: "TLS_DHE_RSA_WITH_AES_256_GCM_SHA384",
-    0x1301: "TLS_AES_128_GCM_SHA256",
-    0x1302: "TLS_AES_256_GCM_SHA384",
-    0x1303: "TLS_CHACHA20_POLY1305_SHA256",
-    0xC008: "TLS_ECDHE_ECDSA_WITH_3DES_EDE_CBC_SHA",
-    0xC009: "TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA",
-    0xC00A: "TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA",
-    0xC012: "TLS_ECDHE_RSA_WITH_3DES_EDE_CBC_SHA",
-    0xC013: "TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA",
-    0xC014: "TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA",
-    0xC023: "TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA256",
-    0xC024: "TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA384",
-    0xC027: "TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256",
-    0xC028: "TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA384",
-    0xC02B: "TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256",
-    0xC02C: "TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384",
-    0xC02F: "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256",
-    0xC030: "TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384",
-    0xC035: "TLS_ECDHE_PSK_WITH_AES_128_CBC_SHA",
-    0xC036: "TLS_ECDHE_PSK_WITH_AES_256_CBC_SHA",
-    0xCCA8: "TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256",
-    0xCCA9: "TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256",
-    0xCCAC: "TLS_ECDHE_PSK_WITH_CHACHA20_POLY1305_SHA256",
-}
+TLS_CIPHER_NAME_MAP = {cipher.value: cipher.name for cipher in CipherName}
 
 # RFC tls extensions: https://datatracker.ietf.org/doc/html/rfc6066
 # IANA list: https://www.iana.org/assignments/tls-extensiontype-values/tls-extensiontype-values.xhtml
@@ -357,8 +369,6 @@ TLS_EXTENSION_NAME_MAP = {
 }
 
 
-
-
 TLS_EC_CURVES_MAP = {
     19: "P-192",
     21: "P-224",
@@ -445,4 +455,3 @@ def toggle_extension(curl, extension_id: int, enable: bool):
             f"This extension({extension_id}) can not be toggled for now, it may be "
             "updated later."
         )
-
