+++ curl_cffi/utils.py
@@ -3,6 +3,7 @@ from typing import Literal
 
 HttpVersionLiteral = Literal["v1", "v2", "v2tls", "v2_prior_knowledge", "v3", "v3only"]
 
+
 class CurlCffiWarning(UserWarning, RuntimeWarning):
     pass
 
@@ -12,4 +13,3 @@ def config_warnings(on: bool = False):
         warnings.simplefilter("default", category=CurlCffiWarning)
     else:
         warnings.simplefilter("ignore", category=CurlCffiWarning)
-
