+++ examples/websockets/simple_echo_async.py
@@ -9,7 +9,6 @@ async def main() -> None:
     async with AsyncSession[Response]() as s:
         print("Connecting to Postman Echo...")
         async with s.ws_connect("wss://ws.postman-echo.com/raw") as ws:
-
             message = "Hello from curl_cffi!"
             print(f">>> Sending: {message}")
             await ws.send_str(payload=message)
